Add adding user

+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ use App\User;
 use App\Role;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
@@ -25,17 +26,47 @@ class UserController extends Controller
         return view('errors.403');
     }
 
+    public function create(Request $request)
+    {
+        if (Auth::user()->can('create', User::class))
+        {
+            User::updateOrCreate(['email' => $request->email], [
+                'name' => $request->name,
+                'password' => Hash::make($request->password),
+                'role_id' => $request->role,
+            ]);
+
+            return redirect()->back();
+        }
+
+        return view('errors.403');
+    }
+
     public function update(Request $request)
     {
-        User::where('id', $request->id)->update(['name' => $request->name, 'email' => $request->email,
-            'role_id' => $request->role]);
-        
-        return redirect()->route('users.view');
+        $userToUpdate = User::where('id', $request->id)->firstOrFail();
+
+        if (Auth::user()->can('update', $userToUpdate))
+        {
+            User::where('id', $request->id)->update(['name' => $request->name, 'email' => $request->email,
+                'role_id' => $request->role]);
+
+            return redirect()->back();
+        }
+
+        return view('errors.403');
     }
 
     public function delete(Request $request)
     {
-        User::where('id', $request->id)->delete();
-        return redirect()->route('users.view');
+        $userToDelete = User::where('id', $request->id)->firstOrFail();
+
+        if (Auth::user()->can('delete', $userToDelete))
+        {
+            User::where('id', $request->id)->delete();
+            return redirect()->back();
+        }
+
+        return view('errors.403');
     }
 }
